[lld][ELF] Allow non-power-of-2 values in ALIGN builtin functions
[1] introduced a requirement that inputs to the ALIGN buitin functions
and the optional ALIGN in an output section description be a power of 2. As
best I can tell, there isn't an external restriction against supporting
non-power-of-2 values in the ALIGN builtin functions and internally these
seem to be handled similarly to any other linker script expression where we
don't have any hard assumptions about the alignment. As allowing non-power-of-2
values aligns with binutils ld and some images (Zephyr, for example) make use
of this, remove the power of 2 requirement for the align builtin functions.

[1] https://github.com/llvm/llvm-project/commit/f22ec9ddf61a1328f77f039c786cab5d79397ffb